First Attempt At Implementing A Timer and GSFU Code To Game Manager
This attempt is using a stopwatch class. I start the stopwatch at the start of the question and pull the elapsed time out from the stopwatch at the question being correct. It isn't done or tested and records in milliseconds, but if it works I can fix everything later.
My name and the table name is hard coded in at the moment. Will fix later.